updating useRecordTable to use internal hooks
