+++ front/src/modules/ui/layout/board/hooks/useBoardColumns.ts
@@ -1,6 +1,6 @@
 import { useRecoilState } from 'recoil';
 
-import { useMoveViewColumns } from '@/ui/object/record-table/hooks/useMoveViewColumns';
+import { useMoveViewColumns } from '@/views/hooks/useMoveViewColumns';
 import { useUpdatePipelineStageMutation } from '~/generated/graphql';
 
 import { boardColumnsState } from '../states/boardColumnsState';
