+++ front/src/modules/ui/object/record-table/hooks/useRecordTable.ts
@@ -1,21 +1,31 @@
 import { useRecoilCallback, useSetRecoilState } from 'recoil';
+import { Key } from 'ts-key-enum';
 
 import { RecordTableScopeInternalContext } from '@/ui/object/record-table/scopes/scope-internal-context/RecordTableScopeInternalContext';
 import { onColumnsChangeScopedState } from '@/ui/object/record-table/states/onColumnsChangeScopedState';
+import { useScopedHotkeys } from '@/ui/utilities/hotkey/hooks/useScopedHotkeys';
+import { useSetHotkeyScope } from '@/ui/utilities/hotkey/hooks/useSetHotkeyScope';
 import { useAvailableScopeIdOrThrow } from '@/ui/utilities/recoil-scope/scopes-internal/hooks/useAvailableScopeId';
 import { getScopedState } from '@/ui/utilities/recoil-scope/utils/getScopedState';
 import { getSnapshotValue } from '@/ui/utilities/recoil-scope/utils/getSnapshotValue';
 
 import { FieldMetadata } from '../../field/types/FieldMetadata';
+import { numberOfTableRowsState } from '../states/numberOfTableRowsState';
 import { onEntityCountChangeScopedState } from '../states/onEntityCountChange';
+import { numberOfTableColumnsScopedSelector } from '../states/selectors/numberOfTableColumnsScopedSelector';
+import { softFocusPositionState } from '../states/softFocusPositionState';
 import { ColumnDefinition } from '../types/ColumnDefinition';
+import { TableHotkeyScope } from '../types/TableHotkeyScope';
 
+import { useDisableSoftFocus } from './internal/useDisableSoftFocus';
 import { useGetIsSomeCellInEditMode } from './internal/useGetIsSomeCellInEditMode';
 import { useLeaveTableFocus } from './internal/useLeaveTableFocus';
 import { useRecordTableScopedStates } from './internal/useRecordTableScopedStates';
 import { useResetTableRowSelection } from './internal/useResetTableRowSelection';
+import { useSelectAllRows } from './internal/useSelectAllRows';
 import { useSetRecordTableData } from './internal/useSetRecordTableData';
 import { useSetRowSelectedState } from './internal/useSetRowSelectedState';
+import { useSetSoftFocusPosition } from './internal/useSetSoftFocusPosition';
 import { useUpsertRecordTableItem } from './internal/useUpsertRecordTableItem';
 
 type useRecordTableProps = {
@@ -93,6 +103,196 @@ export const useRecordTable = (props?: useRecordTableProps) => {
 
   const upsertRecordTableItem = useUpsertRecordTableItem();
 
+  const setSoftFocusPosition = useSetSoftFocusPosition();
+
+  const moveUp = useRecoilCallback(
+    ({ snapshot }) =>
+      () => {
+        const softFocusPosition = snapshot
+          .getLoadable(softFocusPositionState)
+          .valueOrThrow();
+
+        let newRowNumber = softFocusPosition.row - 1;
+
+        if (newRowNumber < 0) {
+          newRowNumber = 0;
+        }
+
+        setSoftFocusPosition({
+          ...softFocusPosition,
+          row: newRowNumber,
+        });
+      },
+    [setSoftFocusPosition],
+  );
+
+  const moveDown = useRecoilCallback(
+    ({ snapshot }) =>
+      () => {
+        const softFocusPosition = snapshot
+          .getLoadable(softFocusPositionState)
+          .valueOrThrow();
+
+        const numberOfTableRows = snapshot
+          .getLoadable(numberOfTableRowsState)
+          .valueOrThrow();
+
+        let newRowNumber = softFocusPosition.row + 1;
+
+        if (newRowNumber >= numberOfTableRows) {
+          newRowNumber = numberOfTableRows - 1;
+        }
+
+        setSoftFocusPosition({
+          ...softFocusPosition,
+          row: newRowNumber,
+        });
+      },
+    [setSoftFocusPosition],
+  );
+
+  const moveRight = useRecoilCallback(
+    ({ snapshot }) =>
+      () => {
+        const softFocusPosition = snapshot
+          .getLoadable(softFocusPositionState)
+          .valueOrThrow();
+
+        const numberOfTableColumns = snapshot
+          .getLoadable(numberOfTableColumnsScopedSelector(scopeId))
+          .valueOrThrow();
+
+        const numberOfTableRows = snapshot
+          .getLoadable(numberOfTableRowsState)
+          .valueOrThrow();
+
+        const currentColumnNumber = softFocusPosition.column;
+        const currentRowNumber = softFocusPosition.row;
+
+        const isLastRowAndLastColumn =
+          currentColumnNumber === numberOfTableColumns - 1 &&
+          currentRowNumber === numberOfTableRows - 1;
+
+        const isLastColumnButNotLastRow =
+          currentColumnNumber === numberOfTableColumns - 1 &&
+          currentRowNumber !== numberOfTableRows - 1;
+
+        const isNotLastColumn =
+          currentColumnNumber !== numberOfTableColumns - 1;
+
+        if (isLastRowAndLastColumn) {
+          return;
+        }
+
+        if (isNotLastColumn) {
+          setSoftFocusPosition({
+            row: currentRowNumber,
+            column: currentColumnNumber + 1,
+          });
+        } else if (isLastColumnButNotLastRow) {
+          setSoftFocusPosition({
+            row: currentRowNumber + 1,
+            column: 0,
+          });
+        }
+      },
+    [scopeId, setSoftFocusPosition],
+  );
+
+  const moveLeft = useRecoilCallback(
+    ({ snapshot }) =>
+      () => {
+        const softFocusPosition = snapshot
+          .getLoadable(softFocusPositionState)
+          .valueOrThrow();
+
+        const numberOfTableColumns = snapshot
+          .getLoadable(numberOfTableColumnsScopedSelector(scopeId))
+          .valueOrThrow();
+
+        const currentColumnNumber = softFocusPosition.column;
+        const currentRowNumber = softFocusPosition.row;
+
+        const isFirstRowAndFirstColumn =
+          currentColumnNumber === 0 && currentRowNumber === 0;
+
+        const isFirstColumnButNotFirstRow =
+          currentColumnNumber === 0 && currentRowNumber > 0;
+
+        const isNotFirstColumn = currentColumnNumber > 0;
+
+        if (isFirstRowAndFirstColumn) {
+          return;
+        }
+
+        if (isNotFirstColumn) {
+          setSoftFocusPosition({
+            row: currentRowNumber,
+            column: currentColumnNumber - 1,
+          });
+        } else if (isFirstColumnButNotFirstRow) {
+          setSoftFocusPosition({
+            row: currentRowNumber - 1,
+            column: numberOfTableColumns - 1,
+          });
+        }
+      },
+    [scopeId, setSoftFocusPosition],
+  );
+
+  const useMapKeyboardToSoftFocus = () => {
+    const disableSoftFocus = useDisableSoftFocus();
+    const setHotkeyScope = useSetHotkeyScope();
+
+    useScopedHotkeys(
+      [Key.ArrowUp, `${Key.Shift}+${Key.Enter}`],
+      () => {
+        moveUp();
+      },
+      TableHotkeyScope.TableSoftFocus,
+      [moveUp],
+    );
+
+    useScopedHotkeys(
+      Key.ArrowDown,
+      () => {
+        moveDown();
+      },
+      TableHotkeyScope.TableSoftFocus,
+      [moveDown],
+    );
+
+    useScopedHotkeys(
+      [Key.ArrowLeft, `${Key.Shift}+${Key.Tab}`],
+      () => {
+        moveLeft();
+      },
+      TableHotkeyScope.TableSoftFocus,
+      [moveLeft],
+    );
+
+    useScopedHotkeys(
+      [Key.ArrowRight, Key.Tab],
+      () => {
+        moveRight();
+      },
+      TableHotkeyScope.TableSoftFocus,
+      [moveRight],
+    );
+
+    useScopedHotkeys(
+      [Key.Escape],
+      () => {
+        setHotkeyScope(TableHotkeyScope.Table, { goto: true });
+        disableSoftFocus();
+      },
+      TableHotkeyScope.TableSoftFocus,
+      [disableSoftFocus],
+    );
+  };
+
+  const { selectAllRows } = useSelectAllRows();
+
   return {
     scopeId,
     onColumnsChange,
@@ -106,5 +306,11 @@ export const useRecordTable = (props?: useRecordTableProps) => {
     setRowSelectedState,
     resetTableRowSelection,
     upsertRecordTableItem,
+    moveDown,
+    moveLeft,
+    moveRight,
+    moveUp,
+    useMapKeyboardToSoftFocus,
+    selectAllRows,
   };
 };
