+++ front/src/modules/ui/object/record-table/hooks/useUpsertTableRowId.ts
@@ -2,6 +2,9 @@ import { useRecoilCallback } from 'recoil';
 
 import { tableRowIdsState } from '../states/tableRowIdsState';
 
+// Used only in company table and people table
+// Remove after refactoring
+
 export const useUpsertTableRowId = () =>
   useRecoilCallback(
     ({ set, snapshot }) =>
